Day 56_57_58_59_60 LLM & RAGs/Game Of Thrones Q&A Bot/requirement.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
streamlit
sentence-transformers
transformers
torch

# use python 3.11 #